+++ src/components/game/cards/CardDetails.css
