+++ src/components/game/cards/DrawnCard.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from 'react';
 import { useSelector } from 'react-redux';
-import { RootState } from '../../store';
+import { RootState } from '../../../store';
 import './DrawnCard.css';
 import CardDetails from './CardDetails';
